acceptance: split util.go
