+++ pkg/acceptance/util_gossip.go
@@ -1,0 +1,83 @@
+// Copyright 2015 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package acceptance
+
+import (
+	"strings"
+	"time"
+
+	"golang.org/x/net/context"
+
+	"github.com/cockroachdb/cockroach/pkg/acceptance/cluster"
+	"github.com/cockroachdb/cockroach/pkg/gossip"
+	"github.com/cockroachdb/cockroach/pkg/util"
+	"github.com/cockroachdb/cockroach/pkg/util/httputil"
+	"github.com/pkg/errors"
+)
+
+// CheckGossipFunc is the type of callback used in CheckGossip.
+type CheckGossipFunc func(map[string]gossip.Info) error
+
+// CheckGossip fetches the gossip infoStore from each node and invokes the given
+// function. The test passes if the function returns 0 for every node,
+// retrying for up to the given duration.
+func CheckGossip(ctx context.Context, c cluster.Cluster, d time.Duration, f CheckGossipFunc) error {
+	return errors.Wrapf(util.RetryForDuration(d, func() error {
+		var infoStatus gossip.InfoStatus
+		for i := 0; i < c.NumNodes(); i++ {
+			if err := httputil.GetJSON(cluster.HTTPClient, c.URL(ctx, i)+"/_status/gossip/local", &infoStatus); err != nil {
+				return errors.Wrapf(err, "failed to get gossip status from node %d", i)
+			}
+			if err := f(infoStatus.Infos); err != nil {
+				return errors.Wrapf(err, "node %d", i)
+			}
+		}
+
+		return nil
+	}), "condition failed to evaluate within %s", d)
+}
+
+// HasPeers returns a CheckGossipFunc that passes when the given
+// number of peers are connected via gossip.
+func HasPeers(expected int) CheckGossipFunc {
+	return func(infos map[string]gossip.Info) error {
+		count := 0
+		for k := range infos {
+			if strings.HasPrefix(k, "node:") {
+				count++
+			}
+		}
+		if count != expected {
+			return errors.Errorf("expected %d peers, found %d", expected, count)
+		}
+		return nil
+	}
+}
+
+// hasSentinel is a checkGossipFunc that passes when the sentinel gossip is present.
+func hasSentinel(infos map[string]gossip.Info) error {
+	if _, ok := infos[gossip.KeySentinel]; !ok {
+		return errors.Errorf("sentinel not found")
+	}
+	return nil
+}
+
+// hasClusterID is a checkGossipFunc that passes when the cluster ID gossip is present.
+func hasClusterID(infos map[string]gossip.Info) error {
+	if _, ok := infos[gossip.KeyClusterID]; !ok {
+		return errors.Errorf("cluster ID not found")
+	}
+	return nil
+}
